Fix CI: mypy strict (79 errors) + cargo fmt
mypy strict had 79 issues, cargo fmt wanted alphabetical module order.

Root causes & fixes:
- scipy.* not in mypy ignore list → added.
- solve_cgp polymorphic return based on return_fields → added @overload
  with Literal[True]/Literal[False] dispatch.
- Convenience funcs in __init__.py passed str to Pydantic models typed as
  float (BeforeValidator coerces at runtime, but mypy sees float-only)
  → added _to_m() helper that calls parse_length() on strings before
  the model constructor.
- numpy bool/array returns flagged 'Returning Any' → added explicit
  float()/bool() casts and typed local annotations.
- Unused type: ignore comments removed.
- CLI: Path/str fields with None default needed Path | None / str | None
  to match Typer's runtime semantics.
- Cargo modules sorted alphabetically.

187/187 tests still pass; ruff, black, cargo test, cargo clippy clean.

+++ src/atlc3/analytical/hammerstad.py
@@ -52,7 +52,7 @@ def _eps_eff_thin_strip(W: float, H: float, er: float) -> float:
         + (1.0 / 18.7) * math.log(1.0 + (u / 18.1) ** 3)
     )
     b = 0.564 * ((er - 0.9) / (er + 3.0)) ** 0.053
-    return (er + 1.0) / 2.0 + ((er - 1.0) / 2.0) * (1.0 + 10.0 / u) ** (-a * b)
+    return float((er + 1.0) / 2.0 + ((er - 1.0) / 2.0) * (1.0 + 10.0 / u) ** (-a * b))
 
 
 def _z0_thin_strip(W: float, H: float, eps_eff: float) -> float:
